% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bcluster}
\alias{bcluster}
\title{Clustering the given objects sequence}
\usage{
bcluster(objects, compares)
}
\arguments{
\item{objects}{a list of data objects}

\item{compares}{evaluate the orders between any element object in \code{objects},
see \code{\link{buildBTree}}.}
}
\value{
a list of list key collection with multiple cluster bins
}
\description{
Clustering the given objects sequence
}
