% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{binaryTree}
\alias{binaryTree}
\title{Create a new tree node}
\usage{
binaryTree(i, ref)
}
\arguments{
\item{i}{the index position of the content data object}

\item{ref}{the reference location of the current tree node}
}
\description{
Create a new tree node
}
\details{
the initial node content data object \code{i} is already includes into the
    \code{members} slot of the created new binary tree node list object.
}
