% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bnetwork}
\alias{bnetwork}
\title{Create a similarity network based on the clustering}
\usage{
bnetwork(objects, compares, show_progress = TRUE)
}
\description{
Create a similarity network based on the clustering
}
