% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{right}
\alias{right}
\title{Get node on current right}
\usage{
right(ref, tree)
}
\arguments{
\item{ref}{the reference location of current node}

\item{tree}{a binary tree list that produced by \code{\link{buildBTree}}.}
}
\value{
a S4 class object of \code{binTree}
}
\description{
Get node on current right
}
