% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{binaryTree}
\alias{binaryTree}
\title{Create a new tree node}
\usage{
binaryTree(i)
}
\arguments{
\item{x}{a node content data object}
}
\description{
Create a new tree node
}
\details{
the initial node content data object \code{x} is already includes into the
    \code{members} slot of the created new binary tree node list object.
}
