% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binTree.R
\name{buildBTree}
\alias{buildBTree}
\title{Build a binary tree}
\usage{
buildBTree(objects, compares, verbose = TRUE)
}
\arguments{
\item{objects}{this parameter should be an object list.}

\item{compares}{a comparision function that compares the order of
any two elements in objects list. this function should generates
3 values for build a tree cluster:
0 for identical, means clustering
1 for great than, means append to right node and
-1 means less than, which will append to left node}
}
\description{
Build a binary tree
}
